Export `WholeShareStockTrade`
